visual_models: move VisualModelForm to Views/, drop _form partial
Components::VisualModelForm -> Views::Controllers::VisualModels::Form
Per #4331. _form.html.erb wrapper deleted; new.html.erb and
edit.html.erb render the Phlex view directly.

Also fixes existing style nits in visual_models/edit.html.erb:
parens on link_to + double-quoted strings (the file was on the
old Rails-y no-parens style).
