--- app/views/controllers/visual_models/_form.html.erb
@@ -1,1 +1,0 @@
-<%= render(Components::VisualModelForm.new(visual_model)) %>
